モグラの再生成
