+++ Assets/Scenes/main.unity
@@ -38,7 +38,7 @@ RenderSettings:
   m_ReflectionIntensity: 1
   m_CustomReflection: {fileID: 0}
   m_Sun: {fileID: 705507994}
-  m_IndirectSpecularColor: {r: 0.6665959, g: 0.88282853, b: 0.885736, a: 1}
+  m_IndirectSpecularColor: {r: 0.66614175, g: 0.88273513, b: 0.88569474, a: 1}
   m_UseRadianceAmbientProbe: 0
 --- !u!157 &3
 LightmapSettings:
@@ -921,6 +921,7 @@ GameObject:
   - component: {fileID: 345832275}
   - component: {fileID: 345832274}
   - component: {fileID: 345832276}
+  - component: {fileID: 345832277}
   m_Layer: 0
   m_Name: GameManager
   m_TagString: Untagged
@@ -972,8 +973,22 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 8de9ee434b41d46d99275483c6438325, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  blockParent: {fileID: 2909061866247334754, guid: b43bad4768d95495482cc752d8dd43d6, type: 3}
+  blockParent: {fileID: 1900529677}
   blockPrefab_Grass: {fileID: 2527857454293643736, guid: b43bad4768d95495482cc752d8dd43d6, type: 3}
+--- !u!114 &345832277
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 345832273}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 51bb318f57f0144afbeec5eb19ac4564, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  blockParent: {fileID: 1573774492}
+  blockPrefab_Mogura: {fileID: 8905744113825350624, guid: 77d3fccc9002042859c4e317b09260bf, type: 3}
 --- !u!1001 &447625068
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -1286,6 +1301,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   ScoreFlag: 0
   newPrefab: {fileID: 3934356072360846589, guid: b961387b7b3a646efb15721f5ac34528, type: 3}
+  isMoguraDestory: 0
 --- !u!4 &553699507 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 4873706082057161953, guid: d8969eedf7d21434690b0c0a27db2b22, type: 3}
@@ -2016,7 +2032,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 788718726}
   serializedVersion: 2
-  m_LocalRotation: {x: -1e-45, y: 1e-45, z: 0, w: 1}
+  m_LocalRotation: {x: 3e-45, y: 1e-45, z: -0, w: 1}
   m_LocalPosition: {x: -0.56, y: 30, z: -111}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
@@ -2283,7 +2299,7 @@ Transform:
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 500, y: 10, z: 500}
+  m_LocalScale: {x: 500, y: 1, z: 500}
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 0}
@@ -2423,7 +2439,7 @@ Camera:
   m_Depth: -1
   m_CullingMask:
     serializedVersion: 2
-    m_Bits: 55
+    m_Bits: 247
   m_RenderingPath: -1
   m_TargetTexture: {fileID: 0}
   m_TargetDisplay: 0
@@ -2443,7 +2459,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 963194225}
   serializedVersion: 2
-  m_LocalRotation: {x: -1e-45, y: 1e-45, z: 0, w: 1}
+  m_LocalRotation: {x: 3e-45, y: 1e-45, z: 0, w: 1}
   m_LocalPosition: {x: -0.56, y: 30, z: -111}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
@@ -3053,7 +3069,7 @@ Camera:
   m_Depth: -2
   m_CullingMask:
     serializedVersion: 2
-    m_Bits: 72
+    m_Bits: 200
   m_RenderingPath: -1
   m_TargetTexture: {fileID: 8400000, guid: 34af762f97b2f45d88d5ce78ea7df677, type: 2}
   m_TargetDisplay: 0
@@ -3641,6 +3657,11 @@ Transform:
 Transform:
   m_CorrespondingSourceObject: {fileID: 4873706082057161953, guid: d8969eedf7d21434690b0c0a27db2b22, type: 3}
   m_PrefabInstance: {fileID: 1672897350}
+  m_PrefabAsset: {fileID: 0}
+--- !u!4 &1573774492 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 8132480282405525850, guid: 77d3fccc9002042859c4e317b09260bf, type: 3}
+  m_PrefabInstance: {fileID: 2034534250}
   m_PrefabAsset: {fileID: 0}
 --- !u!1 &1577014099
 GameObject:
@@ -4310,6 +4331,11 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 4873706082057161953, guid: d8969eedf7d21434690b0c0a27db2b22, type: 3}
   m_PrefabInstance: {fileID: 2137144885}
   m_PrefabAsset: {fileID: 0}
+--- !u!4 &1900529677 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: -8679921383154817045, guid: ecdeebbc0b68f4d7cb7a01ad33d27838, type: 3}
+  m_PrefabInstance: {fileID: 69634656}
+  m_PrefabAsset: {fileID: 0}
 --- !u!1001 &1925753359
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -4634,6 +4660,75 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 4873706082057161953, guid: d8969eedf7d21434690b0c0a27db2b22, type: 3}
   m_PrefabInstance: {fileID: 1880976936}
   m_PrefabAsset: {fileID: 0}
+--- !u!1001 &2034534250
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 8132480282405525850, guid: 77d3fccc9002042859c4e317b09260bf, type: 3}
+      propertyPath: m_LocalScale.x
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 8132480282405525850, guid: 77d3fccc9002042859c4e317b09260bf, type: 3}
+      propertyPath: m_LocalScale.y
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 8132480282405525850, guid: 77d3fccc9002042859c4e317b09260bf, type: 3}
+      propertyPath: m_LocalScale.z
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 8132480282405525850, guid: 77d3fccc9002042859c4e317b09260bf, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8132480282405525850, guid: 77d3fccc9002042859c4e317b09260bf, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: -42.8
+      objectReference: {fileID: 0}
+    - target: {fileID: 8132480282405525850, guid: 77d3fccc9002042859c4e317b09260bf, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8132480282405525850, guid: 77d3fccc9002042859c4e317b09260bf, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 8132480282405525850, guid: 77d3fccc9002042859c4e317b09260bf, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8132480282405525850, guid: 77d3fccc9002042859c4e317b09260bf, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8132480282405525850, guid: 77d3fccc9002042859c4e317b09260bf, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8132480282405525850, guid: 77d3fccc9002042859c4e317b09260bf, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8132480282405525850, guid: 77d3fccc9002042859c4e317b09260bf, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8132480282405525850, guid: 77d3fccc9002042859c4e317b09260bf, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8905744113825350624, guid: 77d3fccc9002042859c4e317b09260bf, type: 3}
+      propertyPath: m_Name
+      value: Mogura
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+    m_RemovedGameObjects: []
+    m_AddedGameObjects: []
+    m_AddedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 77d3fccc9002042859c4e317b09260bf, type: 3}
 --- !u!1 &2046394519
 GameObject:
   m_ObjectHideFlags: 0
@@ -5203,3 +5298,4 @@ SceneRoots:
   - {fileID: 1547051173}
   - {fileID: 788718728}
   - {fileID: 1183952989}
+  - {fileID: 2034534250}
