+++ Assets/Scripts/MoguraManager.cs.meta
