+++ ProjectSettings/TagManager.asset
@@ -18,7 +18,7 @@ TagManager:
   - Water
   - UI
   - Ground
-  - 
+  - Mogura
   - 
   - 
   -
